Fix the nav_bar id to the top and pad; pad the body_content id and adjust both relationally
Make sure that the nav bar where Home link stays is fixed
and make sure that both the nav bar and the rest of the content
interact accordingly. That is, ensuring that when scrolling the content
goes below the nav bar and ensure that before scrolling none of
the header (h1) appears to be underneath the nav bar by use of top padding.